# src/main/java/com/service/support/servicesupport_be_v1/web/api/OwnerCompanyApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.service.support.servicesupport_be_v1.web.api;

import com.service.support.servicesupport_be_v1.web.model.OwnerCompany;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-21T20:52:41.212096500+02:00[Europe/Budapest]")
@Validated
@Tag(name = "owner_company", description = "the owner_company API")
public interface OwnerCompanyApi {

    /**
     * GET /owner-companies : Get all owner companies
     *
     * @return List of companies (status code 200)
     */
    @Operation(
        operationId = "ownerCompaniesGet",
        summary = "Get all owner companies",
        tags = { "owner_company" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of companies", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OwnerCompany.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/owner-companies",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<OwnerCompany>> ownerCompaniesGet(
        
    );


    /**
     * DELETE /owner-companies/{id} : Delete owner company
     *
     * @param id  (required)
     * @return Deleted (status code 204)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "ownerCompaniesIdDelete",
        summary = "Delete owner company",
        tags = { "owner_company" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Deleted"),
            @ApiResponse(responseCode = "404", description = "Not found")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/owner-companies/{id}"
    )
    
    ResponseEntity<Void> ownerCompaniesIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );


    /**
     * GET /owner-companies/{id} : Get owner company by ID
     *
     * @param id  (required)
     * @return Company found (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "ownerCompaniesIdGet",
        summary = "Get owner company by ID",
        tags = { "owner_company" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Company found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OwnerCompany.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/owner-companies/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<OwnerCompany> ownerCompaniesIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );


    /**
     * PUT /owner-companies/{id} : Update owner company
     *
     * @param id  (required)
     * @param ownerCompany  (required)
     * @return Updated (status code 200)
     */
    @Operation(
        operationId = "ownerCompaniesIdPut",
        summary = "Update owner company",
        tags = { "owner_company" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OwnerCompany.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/owner-companies/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<OwnerCompany> ownerCompaniesIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "OwnerCompany", description = "", required = true) @Valid @RequestBody OwnerCompany ownerCompany
    );


    /**
     * POST /owner-companies : Create a new owner company
     *
     * @param ownerCompany  (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "ownerCompaniesPost",
        summary = "Create a new owner company",
        tags = { "owner_company" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OwnerCompany.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/owner-companies",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<OwnerCompany> ownerCompaniesPost(
        @Parameter(name = "OwnerCompany", description = "", required = true) @Valid @RequestBody OwnerCompany ownerCompany
    );

}