# src/main/java/com/service/support/servicesupport_be_v1/web/api/SparePartApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.service.support.servicesupport_be_v1.web.api;

import com.service.support.servicesupport_be_v1.web.model.SparePart;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-21T20:52:41.212096500+02:00[Europe/Budapest]")
@Validated
@Tag(name = "spare_part", description = "the spare_part API")
public interface SparePartApi {

    /**
     * GET /spare-parts : List or search spare parts
     *
     * @param name If provided, filters spare parts by partial name match (optional)
     * @return List of spare parts (status code 200)
     */
    @Operation(
        operationId = "sparePartsGet",
        summary = "List or search spare parts",
        tags = { "spare_part" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of spare parts", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = SparePart.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/spare-parts",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<SparePart>> sparePartsGet(
        @Parameter(name = "name", description = "If provided, filters spare parts by partial name match", in = ParameterIn.QUERY) @Valid @RequestParam(value = "name", required = false) String name
    );


    /**
     * DELETE /spare-parts/{id} : Delete spare part
     *
     * @param id  (required)
     * @return Spare part deleted (status code 204)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "sparePartsIdDelete",
        summary = "Delete spare part",
        tags = { "spare_part" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Spare part deleted"),
            @ApiResponse(responseCode = "404", description = "Not found")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/spare-parts/{id}"
    )
    
    ResponseEntity<Void> sparePartsIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );


    /**
     * GET /spare-parts/{id} : Get spare part by ID
     *
     * @param id  (required)
     * @return Spare part found (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "sparePartsIdGet",
        summary = "Get spare part by ID",
        tags = { "spare_part" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Spare part found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SparePart.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/spare-parts/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<SparePart> sparePartsIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );


    /**
     * PUT /spare-parts/{id} : Update spare part
     *
     * @param id  (required)
     * @param sparePart  (required)
     * @return Spare part updated (status code 200)
     */
    @Operation(
        operationId = "sparePartsIdPut",
        summary = "Update spare part",
        tags = { "spare_part" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Spare part updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SparePart.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/spare-parts/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<SparePart> sparePartsIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "SparePart", description = "", required = true) @Valid @RequestBody SparePart sparePart
    );


    /**
     * POST /spare-parts : Create a new spare part
     *
     * @param sparePart  (required)
     * @return Spare part created (status code 201)
     */
    @Operation(
        operationId = "sparePartsPost",
        summary = "Create a new spare part",
        tags = { "spare_part" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Spare part created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SparePart.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/spare-parts",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<SparePart> sparePartsPost(
        @Parameter(name = "SparePart", description = "", required = true) @Valid @RequestBody SparePart sparePart
    );

}