# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Service Support API
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: [ ]
paths:
...
  /owner-companies:
    get:
      tags:
        - owner_company
      summary: Get all owner companies
      responses:
        '200':
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OwnerCompany'
    post:
      tags:
        - owner_company
      summary: Create a new owner company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerCompany'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerCompany'
  /owner-companies/{id}:
    get:
      tags:
        - owner_company
      summary: Get owner company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Company found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerCompany'
        '404':
          description: Not found
    put:
      tags:
        - owner_company
      summary: Update owner company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerCompany'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerCompany'
    delete:
      tags:
        - owner_company
      summary: Delete owner company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
...
  /spare-parts:
    get:
      tags: [ spare_part ]
      summary: List or search spare parts
      parameters:
        - in: query
          name: name
          required: false
          schema:
            type: string
          description: If provided, filters spare parts by partial name match
      responses:
        '200':
          description: List of spare parts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SparePart'
    post:
      tags: [ spare_part ]
      summary: Create a new spare part
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SparePart'
      responses:
        '201':
          description: Spare part created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparePart'
  /spare-parts/{id}:
    get:
      tags: [ spare_part ]
      summary: Get spare part by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Spare part found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparePart'
        '404':
          description: Not found
    put:
      tags: [ spare_part ]
      summary: Update spare part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SparePart'
      responses:
        '200':
          description: Spare part updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparePart'
    delete:
      tags: [ spare_part ]
      summary: Delete spare part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Spare part deleted
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SparePart:
      type: object
      properties:
        id:
          type: integer
          example: 1
        itemName:
          type: string
          example: "Csapágy 6203"
        itemNumber:
          type: string
          example: "BRG-6203"
        nettoBuyingPrice:
          type: integer
          example: 1200
        nettoSellingPrice:
          type: integer
          example: 1800
...
    OwnerCompany:
      type: object
      properties:
        id:
          type: integer
        accountNumber:
          type: string
        name:
          type: string
        postalCode:
          type: integer
        street:
          type: string
        taxNumber:
          type: string
        town:
          type: string
...